pyproject.toml
[tool.poetry]
name = "azure-django-metrics"
version = "0.1.0"
description = ""
authors = ["Evandro Pomatti <evandro.pomatti@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
azure-monitor-opentelemetry = "1.1.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-instrumentation-django = "^0.42b0"

[tool.poetry.dev-dependencies]
pylint = "^3.0.2"
pylint-django = "^2.5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
